requirements.txt
discord.py@git+https://github.com/hawk-tomy/discord.py
pyyaml
python-socketio
